BDD nodes
The `Node` class represents an individual BDD node.  Pretty basic; it's
just a record containing the variable, value, and low and high pointers.
A leaf node has a value but not a low or high; an internal node has a
low and a high but not a value.  A leaf node also has a special
"variable" that is distinct from any other variable that you could
provide for internal nodes.

BDD nodes will be ordered by using the built-in `compare` function on
the variable objects that the user passes in.  That lets you use
numbers, symbols, whatever.  To provide a custom ordering, you'd have to
create a new type with its own implementation of `compare`.  I'll
probably add the ability to provide a compare function directly when
creating a BDD to make that easier.

A BDD maintains a cache of all of the nodes that have already been
created, so that it's "reduced".  That means that you can compare nodes
using `identical?` — if two nodes are not identical, then they represent
different boolean formulas.