Thread-local "current" BDD
The functions in the previous patch are a bit cumbersome since you have
to pass in a BDD to each one.  This patch adds `with-bdd` and
`with-new-bdd` macros that keep track of a "current" BDD in a
thread-local variable, which lets you leave out all of the repetitive
`bdd` parameters.