Comparing BDD nodes
This patch adds a `Comparator` implementation for BDD nodes.  Internal
nodes are sorted by their variables, so that every path from the root to
a leaf in a BDD tree encounters variables in the same order.  Leaf nodes
always sort "larger" than any internal node, and equal to other leaves.