all-where
The `all-where` function returns a lazy sequence of variable assignments
that cause a boolean formula to evaluate to a particular value.  The
assignments only include those variables that must have a particular
value; if a variable is missing from an assignment, then it can have any
value for the overall formula to evaluate to the requested value.